Move buildShortcuts() to XmppConnectionService Caching for rounded avatars Refactoring for avatarService
